partial implement move render function
