+++ kernel/src/font_renderer.rs
@@ -30,32 +30,6 @@ lazy_static! {
     };
 }
 
-/// Renders the given character in the given buffer.
-///
-/// This function doesn't support the unicode table since it is only meant for kernel logging and besides, the character cache is too small for unicode.
-fn render_character(character: char, buffer: &mut [u8]) {
-    let glyph_bytes = get_glyph_bitmap(character);
-    let (character_width, character_height) = get_character_dimensions();
-    let bytes_per_row = (character_width + 7) / 8;
-    let bytes_per_pixel = get_bytes_per_pixel();
-    let (red_byte, green_byte, blue_byte) = get_rgb_byte_positions();
-    for glyph_y in 0..character_height {
-        for glyph_x in 0..character_width {
-            let byte = glyph_bytes[glyph_y * bytes_per_row + glyph_x / 8];
-            // Since the character is stored most-significant-bit-first, we have to do this little bit of math to convert our index into a bit index.
-            let bit_index = 7 - glyph_x % 8;
-            let bit = byte & (1 << bit_index) != 0;
-            // We will just fill the buffer with a uniform color (all color bytes the same).
-            // This is a bit limiting, but I think it is ok.
-            let color = if bit { 0xFF } else { 0x00 };
-            let pixel_index = (glyph_y * character_width + glyph_x) * bytes_per_pixel;
-            buffer[pixel_index + red_byte as usize] = color;
-            buffer[pixel_index + green_byte as usize] = color;
-            buffer[pixel_index + blue_byte as usize] = color;
-        }
-    }
-}
-
 pub fn draw_character(character: char, x: usize, y: usize) {
     let (character_width, character_height) = get_character_dimensions();
     let (screen_width, screen_height) = get_screen_dimensions();
